pyproject.toml
[project]
name = "rag-youtube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "click>=8.3.1",
    "cohere>=5.20.4",
    "fastapi>=0.128.6",
    "firecrawl>=4.14.0",
    "langchain>=1.2.9",
    "langchain-community>=0.3.0",
    "langchain-experimental>=0.4.1",
    "langgraph>=1.0.8",
    "nltk>=3.9.2",
    "numpy>=2.4.2",
    "openai>=2.18.0",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "rich>=14.3.2",
    "sentence-transformers>=5.2.2",
    "spacy>=4.0.0.dev3",
    "sqlalchemy>=2.1.0b1",
    "structlog>=25.5.0",
    "supabase>=2.28.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.1.0",
    "uvicorn>=0.40.0",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.2.9.233747.dev0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 120
extend-ignore = ["E501"]

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true
mypy_path = "."
namespace_packages = true
disable_error_code = [
  "var-annotated",
  "attr-defined",
  "union-attr",
  "arg-type",
  "operator",
  "list-item",
  "assignment",
  "no-redef",
  "override",
  "return-value",
  "call-arg",
  "index",
  "import-untyped",
]
